ODROIDC2: fix compile error 'initialization from incompatible pointer type'
/home/tobetter/projects/odroid/c2/android/hardware/wifi/realtek/drivers/8192cu/rtl8xxx_CU/os_dep/linux/rtw_proc.c:324:2:
error: initialization from incompatible pointer type [-Werror]
  {"rx_info", proc_get_rx_info, proc_reset_rx_info},
    ^
    /home/tobetter/projects/odroid/c2/android/hardware/wifi/realtek/drivers/8192cu/rtl8xxx_CU/os_dep/linux/rtw_proc.c:324:2:
    error: (near initialization for ‘adapter_proc_hdls[14].write’) [-Werror]

+++ rtl8xxx_CU/os_dep/linux/rtw_proc.c
@@ -263,7 +263,7 @@ static int proc_get_rf_reg_dump(struct seq_file *m, void *v)
 }
 
 
-int proc_reset_rx_info(struct file *file, const char __user *buffer, size_t count, loff_t *pos, void *data)
+ssize_t proc_reset_rx_info(struct file *file, const char __user *buffer, size_t count, loff_t *pos, void *data)
 {
 	struct net_device *dev = data;
 	_adapter *padapter = (_adapter *)rtw_netdev_priv(dev);
